all working, minor problems
