+++ app/ideciclo/[id]/page.tsx
@@ -4,34 +4,74 @@ import { VerticalStatisticsBoxes } from "../../components/VerticalStatisticsBoxe
 import { StructureMap, StructureRadar } from "./useclient";
 import {
   IDECICLO_FORMS_DATA,
+  IDECICLO_PAGE_DATA,
   IDECICLO_STRUCTURES_DATA,
 } from "../../../servers";
+import { IdecicloDescription } from "./IdecicloDescription";
+import {
+  getGeneralStatistics,
+  get_map_data,
+  rates_organization,
+} from "./configuration";
+import { NavCover } from "../../components/NavCover";
+import { Breadcrumb } from "../../components/Breadcrumb";
 
-export async function getStaticPaths() {
-  const res = await fetch(IDECICLO_STRUCTURES_DATA);
-  const allstructs = await res.json();
-  const rec_structs = allstructs.filter((s) => s.city_id === 1);
-  // Get the paths we want to pre-render based on posts
-  const paths = rec_structs.map((s: any) => ({
-    params: { ideciclo: s.id },
-  }));
+const fetchUniqueData = async (id: string) => {
+  const res = await fetch(IDECICLO_STRUCTURES_DATA + "/" + id);
+  const data = await res.json();
+  return data;
+};
 
-  return { paths, fallback: false };
-}
+// export async function getStaticPaths() {
+//   const res = await fetch(IDECICLO_STRUCTURES_DATA);
+//   const allstructs = await res.json();
+//   const rec_structs = allstructs.filter((s) => s.city_id === 1);
+//   // Get the paths we want to pre-render based on posts
+//   const paths = rec_structs.map((s: any) => ({
+//     params: { ideciclo: s.id },
+//   }));
 
-export async function getStaticProps({ params }) {
-  const res = await fetch(IDECICLO_STRUCTURES_DATA + "/" + params.ideciclo);
+//   return { paths, fallback: false };
+// }
 
-  let structure = await res.json();
-  // iterar sobre reviews
-  // map ou foreach
-  // cada review
+// export async function getStaticProps({ params }) {
+//   const res = await fetch(IDECICLO_STRUCTURES_DATA + "/" + params.ideciclo);
 
+//   let structure = await res.json();
+//   // iterar sobre reviews
+//   // map ou foreach
+//   // cada review
+
+//   /// COLOCAR EM REVIEW NO Back
+//   const new_review_form_id =
+//     structure.reviews[structure.reviews.length - 1].segments[0].form_id;
+
+//   const formres = await fetch(IDECICLO_FORMS_DATA + "/" + new_review_form_id);
+
+//   const forms = await formres.json();
+
+//   let last_review_form_id = null;
+//   if (structure.reviews.length > 1)
+//     last_review_form_id =
+//       structure.reviews[structure.reviews.length - 2].segments[0].form_id;
+
+//   return {
+//     props: {
+//       structure: structure,
+//       forms: forms,
+//     },
+//     revalidate: 1,
+//   };
+// }
+
+const fetchData = async (id, structure) => {
   /// COLOCAR EM REVIEW NO Back
   const new_review_form_id =
     structure.reviews[structure.reviews.length - 1].segments[0].form_id;
 
-  const formres = await fetch(IDECICLO_FORMS_DATA + "/" + new_review_form_id);
+  const formres = await fetch(IDECICLO_FORMS_DATA + "/" + new_review_form_id, {
+    cache: "no-cache",
+  });
 
   const forms = await formres.json();
 
@@ -40,94 +80,54 @@ export async function getStaticProps({ params }) {
     last_review_form_id =
       structure.reviews[structure.reviews.length - 2].segments[0].form_id;
 
-  return {
-    props: {
-      structure: structure,
-      forms: forms,
-    },
-    revalidate: 1,
-  };
-}
+  const pageDataRes = await fetch(IDECICLO_PAGE_DATA, { cache: "no-cache" });
+  const pageData = await pageDataRes.json();
 
-const Ideciclo = ({ structure, forms }) => {
+  return { forms, pageData };
+};
+
+const Ideciclo = async ({ params }) => {
+  const data = await fetchUniqueData(params.id);
+  const { forms, pageData } = await fetchData(params.id, data);
   const page_data = {
-    title: structure.street,
-    cover_image_url: "",
+    title: data.street,
+    src: pageData.cover.url,
   };
 
   const crumb = {
-    label: structure.street,
-    slug: structure.id.toString(),
-    routes: ["/", "/ideciclo", structure.id],
+    label: data.street,
+    slug: data.id.toString(),
+    routes: ["/", "/ideciclo", data.id],
   };
-  let info = rates_organization(structure, forms);
 
-  info.map = get_map_data(structure);
-  const GeneralStatistics = getGeneralStatistics(structure, info);
+  const info = rates_organization(data, forms);
+  const mapData = get_map_data(data);
+  const GeneralStatistics = getGeneralStatistics(data, info);
 
   return (
     <>
+      <NavCover {...page_data} />
+      <Breadcrumb {...crumb} />
       <StatisticsBox
         title={GeneralStatistics.title}
         subtitle={GeneralStatistics.subtitle}
         boxes={GeneralStatistics.boxes}
       />
-
       <section className="container mx-auto grid lg:grid-cols-3 md:grid-cols-1 auto-rows-auto gap-10 my-10">
         <div className="rounded shadow-2xl">
-          <div className="flex flex-col bg-white mx-4 md:mx-auto max-w-4xl divide-y md:divide-x divide-gray-100">
-            <div className="flex flex-col justify-center w-full p-6 text-center tracking-widest">
-              <h3>DESCRIÇÃO</h3>
-              <h3 className="text-2xl mt-2">
-                <strong>{info.tipologia.toUpperCase()}</strong>,{" "}
-                <strong>{info.fluxo.toUpperCase()}</strong>
-                {info.pavimento != null && (
-                  <>
-                    , com piso de{" "}
-                    <strong>
-                      {info.pavimento.replace(",", " e").toUpperCase()}
-                    </strong>
-                  </>
-                )}
-                {info.tipologia.toUpperCase() != "CICLORROTA" && (
-                  <>
-                    , localizada{" "}
-                    <strong>{info.localizacao.toUpperCase()}</strong>
-                  </>
-                )}
-              </h3>
-            </div>
-            <div className="flex flex-col justify-center w-full p-6 text-center tracking-widest">
-              <h3>LARGURA</h3>
-              {info.largura_transitavel >= 0 ? (
-                <h3 className="text-3xl  mt-2">
-                  <strong>
-                    {("" + info.largura_total).replace(".", ",")}m
-                  </strong>
-                  , onde{" "}
-                  <strong>
-                    {("" + info.largura_transitavel).replace(".", ",")}m{" "}
-                  </strong>
-                  são transitáveis
-                </h3>
-              ) : (
-                <h3 className="text-3xl font-bold mt-2">N/A</h3>
-              )}
-            </div>
-            <div className="flex flex-col justify-center w-full p-6 text-center uppercase tracking-widest">
-              <h3>Última avaliação</h3>
-              <h3 className="text-3xl font-bold mt-2">{info.data}</h3>
-            </div>
-          </div>
+          <IdecicloDescription info={info} />
         </div>
         <div className="bg-green-200 rounded shadow-2xl">
-          <StructureMap />
+          <StructureMap data={mapData} />
         </div>
         <div className="rounded shadow-2xl">
-          <StructureRadar />
+          <StructureRadar {...info} />
         </div>
       </section>
-      <VerticalStatisticsBoxes boxes={info} />
+      <VerticalStatisticsBoxes
+        title={"Detalhamento e composição das notas"}
+        boxes={info.parametros}
+      />
     </>
   );
 };
